+++ cmd/xfon/command/x509/x509.go
@@ -1,6 +1,12 @@
 package x509
 
-import "github.com/spf13/cobra"
+import (
+	"log"
+	"os"
+
+	"github.com/odacremolbap/xfon/pkg/cert"
+	"github.com/spf13/cobra"
+)
 
 // RootCmd contains certificate management commands
 var RootCmd = &cobra.Command{
@@ -13,7 +19,7 @@ var RootCmd = &cobra.Command{
 var NewCmd = &cobra.Command{
 	Use:   "new",
 	Short: "creates certificate key pair",
-	Run:   runHelp,
+	Run:   newFunc,
 }
 
 // SignCmd manages private keys
@@ -30,4 +36,33 @@ func runHelp(cmd *cobra.Command, args []string) {
 func init() {
 	RootCmd.AddCommand(NewCmd)
 	RootCmd.AddCommand(SignCmd)
+}
+
+// newFunc runs the new RSA command
+func newFunc(cmd *cobra.Command, args []string) {
+	// Generate certificate
+	// get pem encoded key
+	// get pem encoded public
+	// write key
+	// write certificate
+
+	x := &cert.X509{}
+	err := cert.GenerateX509SelfSignedCertificate(x, nil)
+	if err != nil {
+		log.Printf("error generating RSA key: %v", err.Error())
+		os.Exit(-1)
+	}
+	log.Printf("debugging: %v", x)
+
+	// p, err := cert.RSAtoPEM(k)
+	// if err != nil {
+	// 	log.Printf("error serializing RSA key into PEM: %v", err.Error())
+	// 	os.Exit(-1)
+	// }
+
+	// err = writeCertFile(out, p)
+	// if err != nil {
+	// 	log.Printf("error writing RSA key to file: %v", err.Error())
+	// 	os.Exit(-1)
+	// }
 }
