+++ pkg/cert/cert.go
@@ -1,68 +1,9 @@
 package cert
 
 import (
-	"bytes"
 	"crypto/rand"
-	"crypto/rsa"
-	"crypto/x509"
-	"encoding/pem"
-	"fmt"
 	"math/big"
-	"time"
 )
-
-var (
-	// KeyUsageChoices is a set of string choices that map to the
-	// X509 key usage representation
-	KeyUsageChoices map[string]x509.KeyUsage
-
-	// ExtKeyUsageChoices is a set of string choices that map to the
-	// X509 extended key usage representation
-	ExtKeyUsageChoices map[string]x509.ExtKeyUsage
-)
-
-func init() {
-	KeyUsageChoices = map[string]x509.KeyUsage{
-		"KeyUsageDigitalSignature":  x509.KeyUsageDigitalSignature,
-		"KeyUsageContentCommitment": x509.KeyUsageContentCommitment,
-		"KeyUsageKeyEncipherment":   x509.KeyUsageKeyEncipherment,
-		"KeyUsageDataEncipherment":  x509.KeyUsageDataEncipherment,
-		"KeyUsageKeyAgreement":      x509.KeyUsageKeyAgreement,
-		"KeyUsageCertSign":          x509.KeyUsageCertSign,
-		"KeyUsageCRLSign":           x509.KeyUsageCRLSign,
-		"KeyUsageEncipherOnly":      x509.KeyUsageEncipherOnly,
-		"KeyUsageDecipherOnly":      x509.KeyUsageDecipherOnly,
-	}
-
-	ExtKeyUsageChoices = map[string]x509.ExtKeyUsage{
-		"ExtKeyUsageAny":                            x509.ExtKeyUsageAny,
-		"ExtKeyUsageServerAuth":                     x509.ExtKeyUsageServerAuth,
-		"ExtKeyUsageClientAuth":                     x509.ExtKeyUsageClientAuth,
-		"ExtKeyUsageCodeSigning":                    x509.ExtKeyUsageCodeSigning,
-		"ExtKeyUsageEmailProtection":                x509.ExtKeyUsageEmailProtection,
-		"ExtKeyUsageIPSECEndSystem":                 x509.ExtKeyUsageIPSECEndSystem,
-		"ExtKeyUsageIPSECTunnel":                    x509.ExtKeyUsageIPSECTunnel,
-		"ExtKeyUsageIPSECUser":                      x509.ExtKeyUsageIPSECUser,
-		"ExtKeyUsageTimeStamping":                   x509.ExtKeyUsageTimeStamping,
-		"ExtKeyUsageOCSPSigning":                    x509.ExtKeyUsageOCSPSigning,
-		"ExtKeyUsageMicrosoftServerGatedCrypto":     x509.ExtKeyUsageMicrosoftServerGatedCrypto,
-		"ExtKeyUsageNetscapeServerGatedCrypto":      x509.ExtKeyUsageNetscapeServerGatedCrypto,
-		"ExtKeyUsageMicrosoftCommercialCodeSigning": x509.ExtKeyUsageMicrosoftCommercialCodeSigning,
-		"ExtKeyUsageMicrosoftKernelCodeSigning":     x509.ExtKeyUsageMicrosoftKernelCodeSigning,
-	}
-}
-
-// X509 simplified
-type X509 struct {
-	Serial      *big.Int
-	NotBefore   time.Time
-	NotAfter    time.Time
-	DNSNames    []string
-	IPAddresses []string
-	IsCA        bool
-	KeyUsage    x509.KeyUsage
-	ExtKeyUsage x509.ExtKeyUsage
-}
 
 // Manager manages certificates
 type Manager struct {
@@ -76,25 +17,4 @@ type Manager struct {
 
 func genSerial() (*big.Int, error) {
 	return rand.Int(rand.Reader, (&big.Int{}).Exp(big.NewInt(2), big.NewInt(159), nil))
-}
-
-// GenerateRSAKey using the informed size
-func GenerateRSAKey(bits int) (*rsa.PrivateKey, error) {
-	return rsa.GenerateKey(rand.Reader, bits)
-}
-
-// RSAtoPEM serializes the RSA key into PEM format
-func RSAtoPEM(key *rsa.PrivateKey) (string, error) {
-	var b bytes.Buffer
-	err := pem.Encode(
-		&b,
-		&pem.Block{
-			Type:  "RSA PRIVATE KEY",
-			Bytes: x509.MarshalPKCS1PrivateKey(key),
-		})
-	if err != nil {
-		return "", fmt.Errorf("error PEM encoding RSA key: %s", err.Error())
-	}
-
-	return b.String(), nil
 }
